Update deletion process in cache.js
The code changes improve the deletion process from the kanmi_records_cdn database. Instead of deleting items one by one, we've replaced it with a more efficient batch deletion process. This new approach collects all eids that need to be removed and deletes them together, improving performance especially when dealing with large numbers of items to be deleted.